requirements.txt
fastapi>=0.78.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.9.1,<2.0.0
sqlalchemy>= 1.4.39
alembic>=1.8.1
psycopg2-binary>=2.9.5
pre-commit==3.2.2
python-dotenv==0.21.1
python-multipart==0.0.6